feat: add hook for backend toggle and store for atoms
