+++ src/Modules/DashboardHub.tsx
@@ -1,18 +1,26 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import Header from '../Components/DashboardHub/Header/Header';
 import DashboardTable from '../Components/DashboardHub/DashboardTable/DashboardTable';
 import { PageSection } from '@patternfly/react-core';
 import useGetDashboards from '../hooks/useGetDashboards';
 import Portal from '@redhat-cloud-services/frontend-components-notifications/Portal';
-import { useAtomValue } from 'jotai';
+import { Provider, useAtomValue, useSetAtom } from 'jotai';
+import { useFlag } from '@unleash/proxy-client-react';
 import { notificationsAtom, useRemoveNotification } from '../state/notificationsAtom';
 import { Route, Routes } from 'react-router-dom';
 import GenericDashboardPage from './GenericDashboardPage';
+import { backendFlagAtom, store } from '../state/store';
 
-const DashboardHub = () => {
+const DashboardHubInner = () => {
   const notifications = useAtomValue(notificationsAtom);
   const removeNotification = useRemoveNotification();
   const { dashboards, fetchDashboards } = useGetDashboards();
+  const setBackendFlag = useSetAtom(backendFlagAtom);
+  const isNewBackend = useFlag('platform.widget-layout.new-backend');
+
+  useEffect(() => {
+    setBackendFlag(isNewBackend);
+  }, [isNewBackend]);
 
   return (
     <Routes>
@@ -33,5 +41,11 @@ const DashboardHub = () => {
     </Routes>
   );
 };
+
+const DashboardHub = () => (
+  <Provider store={store}>
+    <DashboardHubInner />
+  </Provider>
+);
 
 export default DashboardHub;
