+++ src/Routes/Default/Default.tsx
@@ -1,7 +1,9 @@
 import { PageSection } from '@patternfly/react-core';
 import AddWidgetDrawer from '../../Components/WidgetDrawer/WidgetDrawer';
 import GridLayout from '../../Components/DnDLayout/GridLayout';
-import { useAtomValue, useSetAtom } from 'jotai';
+import { Provider, useAtomValue, useSetAtom } from 'jotai';
+import { useFlag } from '@unleash/proxy-client-react';
+import { backendFlagAtom, store } from '../../state/store';
 import { lockedLayoutAtom } from '../../state/lockedLayoutAtom';
 import { notificationsAtom, useRemoveNotification } from '../../state/notificationsAtom';
 import Header from '../../Components/Header/Header';
@@ -13,13 +15,19 @@ import Portal from '@redhat-cloud-services/frontend-components-notifications/Por
 import useChrome from '@redhat-cloud-services/frontend-components/useChrome';
 import { resolvedWidgetMappingAtom } from '../../state/widgetMappingAtom';
 
-const DefaultRoute = (props: { layoutType?: LayoutTypes }) => {
+const DefaultRouteInner = (props: { layoutType?: LayoutTypes }) => {
   const isLayoutLocked = useAtomValue(lockedLayoutAtom);
   const notifications = useAtomValue(notificationsAtom);
   const removeNotification = useRemoveNotification();
   const { template, saveTemplate, isLoaded, layoutRef } = useDashboardConfig(props.layoutType);
   const resolveWidgetMapping = useSetAtom(resolvedWidgetMappingAtom);
   const { visibilityFunctions } = useChrome();
+  const setBackendFlag = useSetAtom(backendFlagAtom);
+  const isNewBackend = useFlag('platform.widget-layout.new-backend');
+
+  useEffect(() => {
+    setBackendFlag(isNewBackend);
+  }, [isNewBackend]);
 
   useEffect(() => {
     if (visibilityFunctions) {
@@ -39,5 +47,11 @@ const DefaultRoute = (props: { layoutType?: LayoutTypes }) => {
     </div>
   );
 };
+
+const DefaultRoute = (props: { layoutType?: LayoutTypes }) => (
+  <Provider store={store}>
+    <DefaultRouteInner {...props} />
+  </Provider>
+);
 
 export default DefaultRoute;
