+++ src/state/dashboardsAtom.ts
@@ -1,25 +1,20 @@
 import { atom } from 'jotai';
-import {
-  DashboardTemplate,
-  TemplateConfig,
-  copyDashboardTemplate,
-  deleteDashboardTemplateFromHub,
-  getUsersDashboards,
-  importDashboardTemplate,
-  setDefaultTemplate,
-} from '../api/dashboard-templates';
+import { DashboardTemplate, TemplateConfig } from '../api/dashboard-templates';
+import { getApi } from './store';
 
 export const dashboardsAtom = atom<DashboardTemplate[]>([]);
 
 export const deleteDashboardAtom = atom(null, async (_get, set, id: DashboardTemplate['id']) => {
-  await deleteDashboardTemplateFromHub(id);
-  const dashboards = await getUsersDashboards();
+  const api = getApi();
+  await api.deleteDashboardTemplateFromHub(id);
+  const dashboards = await api.getUsersDashboards();
   set(dashboardsAtom, dashboards);
 });
 
 export const setDefaultDashboardAtom = atom(null, async (_get, set, id: DashboardTemplate['id']) => {
-  await setDefaultTemplate(id);
-  const dashboards = await getUsersDashboards();
+  const api = getApi();
+  await api.setDefaultTemplate(id);
+  const dashboards = await api.getUsersDashboards();
   set(dashboardsAtom, dashboards);
 });
 
@@ -30,12 +25,13 @@ export const createDashboardAtom = atom(
     set,
     data: { dashboardName: string; templateBase: { name: string; displayName: string }; templateConfig: TemplateConfig; setAsHomepage?: boolean }
   ) => {
+    const api = getApi();
     const { setAsHomepage, ...importData } = data;
-    const result = await importDashboardTemplate(importData);
+    const result = await api.importDashboardTemplate(importData);
     if (setAsHomepage) {
-      await setDefaultTemplate(result.id);
+      await api.setDefaultTemplate(result.id);
     }
-    const dashboards = await getUsersDashboards();
+    const dashboards = await api.getUsersDashboards();
     set(dashboardsAtom, dashboards);
     return result;
   }
@@ -44,8 +40,9 @@ export const createDashboardAtom = atom(
 export const importDashboardAtom = atom(
   null,
   async (_get, set, data: { dashboardName: string; templateBase: { name: string; displayName: string }; templateConfig: TemplateConfig }) => {
-    const result = await importDashboardTemplate(data);
-    const dashboards = await getUsersDashboards();
+    const api = getApi();
+    const result = await api.importDashboardTemplate(data);
+    const dashboards = await api.getUsersDashboards();
     set(dashboardsAtom, dashboards);
     return result;
   }
@@ -54,11 +51,12 @@ export const importDashboardAtom = atom(
 export const duplicateDashboardAtom = atom(
   null,
   async (_get, set, data: { id: DashboardTemplate['id']; dashboardName: string; setAsHomepage?: boolean }) => {
-    const result = await copyDashboardTemplate(data.id, { dashboardName: data.dashboardName });
+    const api = getApi();
+    const result = await api.copyDashboardTemplate(data.id, { dashboardName: data.dashboardName });
     if (data.setAsHomepage) {
-      await setDefaultTemplate(result.id);
+      await api.setDefaultTemplate(result.id);
     }
-    const dashboards = await getUsersDashboards();
+    const dashboards = await api.getUsersDashboards();
     set(dashboardsAtom, dashboards);
     return result;
   }
